+++ apps/docs/src/examples/switch/SwitchExample.tsx
@@ -1,0 +1,10 @@
+import { ExampleContainer } from '@/components/ExampleContainer'
+import { Switch } from '@pikas-ui/switch'
+
+export const SwitchExample: React.FC = () => {
+  return (
+    <ExampleContainer>
+      <Switch label="Switch label" id="switch" />
+    </ExampleContainer>
+  )
+}