+++ packages/alert/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @pikas-ui/alert
 
+## 2.0.12
+
+### Patch Changes
+
+- Multiple fix
+- Updated dependencies
+  - @pikas-ui/icons@2.0.12
+  - @pikas-ui/styles@2.2.1
+
 ## 2.0.11
 
 ### Patch Changes