+++ packages/context-menu/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @pikas-ui/context-menu
 
+## 2.0.13
+
+### Patch Changes
+
+- Multiple fix
+- Updated dependencies
+  - @pikas-ui/icons@2.0.12
+  - @pikas-ui/loader@2.0.11
+  - @pikas-ui/menu@2.0.12
+  - @pikas-ui/styles@2.2.1
+
 ## 2.0.12
 
 ### Patch Changes