+++ packages/loader/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @pikas-ui/loader
 
+## 2.0.11
+
+### Patch Changes
+
+- Multiple fix
+- Updated dependencies
+  - @pikas-ui/styles@2.2.1
+
 ## 2.0.10
 
 ### Patch Changes