+++ packages/separator/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @pikas-ui/separator
 
+## 2.0.10
+
+### Patch Changes
+
+- Multiple fix
+- Updated dependencies
+  - @pikas-ui/styles@2.2.1
+
 ## 2.0.9
 
 ### Patch Changes