+++ packages/textarea/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @pikas-ui/textarea
 
+## 2.0.16
+
+### Patch Changes
+
+- Multiple fix
+- Updated dependencies
+  - @pikas-ui/icons@2.0.12
+  - @pikas-ui/loader@2.0.11
+  - @pikas-ui/styles@2.2.1
+  - @pikas-ui/text@3.0.4
+  - @pikas-ui/tooltip@2.0.10
+
 ## 2.0.15
 
 ### Patch Changes